src/api-sk/requirements.txt
fastapi[standard]==0.115.6
aiohttp==3.11.9
pydantic==2.9.2
pydantic-settings==2.6.1
nest_asyncio==1.6.0
opentelemetry-instrumentation-fastapi==0.49b0
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-distro==0.49b0
azure-monitor-opentelemetry==1.6.4
fastapi-profiler==1.4.1
azure-identity==1.19.0
aiofiles==24.1.0
semantic-kernel[azure]==1.16.0